initial Transaction Page
