+++ src/components/Forms/Transactions/TransactionForm.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Customer, Transaction } from "@/types/types";
+import { Transaction } from "@/types/types";
 import AccountBalanceWalletOutlinedIcon from '@mui/icons-material/AccountBalanceWalletOutlined';
 import ArrowForwardIosSharpIcon from '@mui/icons-material/ArrowForwardIosSharp';
 import DetailsIcon from '@mui/icons-material/Details';
@@ -16,8 +16,7 @@ import GeneralTxnFormField from "./GeneralTxnFormField";
 import { LicensingDetailFormFields } from "./LicensingDetailFormFields";
 import { useEffect, useState } from "react";
 import { Autocomplete, Box, TextField } from "@mui/material";
-import { useOne, useTable } from "@refinedev/core";
-import { TURBO_TRACE_DEFAULT_MEMORY_LIMIT } from "next/dist/shared/lib/constants";
+import CustomerForm from "./CustomerForm";
 
 export type TransactionFormProps = GenericFormProps & {
   transaction?: Transaction;
@@ -25,132 +24,114 @@ export type TransactionFormProps = GenericFormProps & {
   reset?: any
 };
 
-const ADDLABEL = "New Customer";
+const TransactionForm = (props: TransactionFormProps) => {
+  const [expandedPanels, setExpandedPanels] = useState<Record<string, boolean>>({ "Transaction": true });
 
-const TransactionForm = ({ reset, ...props }: TransactionFormProps) => {
+  // useEffect(() => {
+  //   const errors = props.errors;
+  //   const newExpandedPanels = FormGroups.reduce((acc, group, index) => {
+  //     const hasErrors = group.fields.some((field: any) => errors?.[field.name]);;
+  //     const shouldExpand = hasErrors || index === 0;
+  //     return { ...acc, [group.title]: shouldExpand };
+  //   }, {});
+  //   setExpandedPanels(newExpandedPanels);
+  // }, [props]);
 
-  const [value, setValue] = useState<any>({
-    isAdd: true,
-    account: ADDLABEL
-  });
-  const [inputValue, setInputValue] = useState(ADDLABEL);
-  const [customers, setCustomers] = useState<any[]>([]);
-  const {
-    tableQueryResult: { data, isLoading },
-    setFilters,
-  } = useTable<Customer>({
-    resource: "customers/bill-customers",
-    syncWithLocation: false
-  });
+  const handleAccordionChange = (panel: string) => {
+    setExpandedPanels((prev) => ({
+      ...prev,
+      [panel]: !prev[panel],
+    }));
+  };
 
-  const handleValueChange = (event: any, newValue: any) => {
-    setValue(newValue);
-  }
-
-
-  useEffect(() => {
-    if (value?.isAdd) {
-      reset({
-        bill_customer_account: '',
-        bill_customer_name: '',
-        bill_address1: '',
-        bill_address2: '',
-        bill_city: '',
-        bill_state: '',
-        bill_postal_code: '',
-        bill_country: '',
-        bill_contact_first_name: '',
-        bill_contact_last_name: '',
-        bill_contact_phone: '',
-        bill_contact_email: '',
-      })
-    } else {
-      reset({
-        bill_customer_account: value?.account,
-        bill_customer_name: value?.name,
-        bill_address1: value?.contact.address?.address1,
-        bill_address2: value?.contact.address?.address2,
-        bill_city: value?.contact.address?.city,
-        bill_state: value?.contact.address?.state,
-        bill_postal_code: value?.contact.address?.postal_code,
-        bill_country: value?.contact.address?.country,
-        bill_contact_first_name: value?.contact.first_name,
-        bill_contact_last_name: value?.contact.last_name,
-        bill_contact_phone: value?.contact.phone,
-        bill_contact_email: value?.contact.email,
-      });
-    }
-  }, [value])
-
-  const handleSearch = (value: string) => setFilters([{ field: 'searchKey', operator: 'contains', value: value }])
-  const handleInputChange = (event: any, newInputValue: any) => {
-    setInputValue(newInputValue);
-    if (newInputValue == ADDLABEL) {
-      handleSearch('')
-    } else {
-      handleSearch(newInputValue)
-    }
-  }
-
-  useEffect(() => {
-    if (data?.data) {
-      setCustomers(data?.data as any[])
-    }
-  }, [data])
-
+  const FormGroups = [
+    {
+      icon: <PaidOutlinedIcon />,
+      title: 'Transaction',
+      description: 'Setup your Transaction data',
+      fields: props.transaction ?
+        GeneralTxnFormField.EditTransactionForm :
+        (props.transaction_action == "New" ? GeneralTxnFormField.CreateTransactionForm.newAction : GeneralTxnFormField.CreateTransactionForm.notNewAction)
+    },
+    {
+      icon: <AccountBalanceWalletOutlinedIcon />,
+      title: 'Billing Partner Information',
+      description: 'Setup your Billing Partner Information',
+      fields: PartnerFormFields.BillingPartnerInformationFormFields,
+      isCustomer: true,
+      customerType: 'bill-customers',
+    },
+    {
+      icon: <PaidOutlinedIcon />,
+      title: 'Shipping Parter Information',
+      description: 'Setup your Shipping Partner Information',
+      fields: PartnerFormFields.ShippingPartnerInformationFormFields,
+      isCustomer: true,
+      customerType: 'ship-customers',
+    },
+    {
+      icon: <ProductionQuantityLimitsOutlinedIcon />,
+      title: 'Reseller Information',
+      description: 'Setup your Reseller Information',
+      fields: PartnerFormFields.ResellerPartnerInformationFormFields,
+      isCustomer: true,
+      customerType: 'resellers',
+    },
+    // {
+    //   icon: <DetailsIcon />,
+    //   title: 'Licensing Details',
+    //   description: 'Setup your Reseller Information',
+    //   fields: props.transaction_action == "New" ? LicensingDetailFormFields.newAction : LicensingDetailFormFields.notNewAction
+    // },
+  ]
   return (
-    <div className="flex justify-center flex-col gap-4">
-      <Autocomplete
-        id="country-select-demo"
-        value={value}
-        onChange={handleValueChange}
-        sx={{ width: 300 }}
-        inputValue={inputValue}
-        onInputChange={handleInputChange}
-        options={[
-          {
-            isAdd: true,
-            account: ADDLABEL
-          },
-          ...customers
-        ]}
-        loading={isLoading}
-        filterOptions={(options) => options}
-        getOptionLabel={(option) => option.account as string}
-        renderOption={(props, option) => {
-          const { key, ...optionProps } = props;
-          return (
-            <Box
-              key={key}
-              component="li"
-              {...optionProps}
+    <div className="flex justify-center">
+      <div className="w-2/3 flex flex-col gap-2">
+        {
+          FormGroups.map((formGroup, i) => (
+            <Accordion
+              key={formGroup.title}
+              expanded={expandedPanels[formGroup.title] || false}
+              onChange={() => handleAccordionChange(formGroup.title)}
+              sx={{
+                borderTop: '2px solid #1f325c',
+                '&::before': { display: 'none' },
+                borderRadius: '0.5rem',
+                boxShadow: '0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1);'
+              }}
             >
-              {
-                option.isAdd == true ?
-                  <div className="border-b">{option.account}</div> :
-                  <div className="text-sm flex flex-col gap-2 py-2">
-                    <div className="flex gap-4">
-                      <div className="border-b">{option.account}</div>
-                      <div className="border-b">{option.contact.first_name + " " + option.contact.last_name}</div>
-                    </div>
-                    <div className="flex gap-2">
-                      <div className="border-b">{option.contact.address?.address1}</div>
-                      <div className="border-b">{option.contact.address?.address2}</div>
-                      <div className="border-b">{option.contact.address?.state}</div>
-                      <div className="border-b">{option.contact.address?.country}</div>
-                    </div>
-                  </div>
-              }
-            </Box>
-          );
-        }}
-        renderInput={(params) => {
-          return (
-            <TextField {...params} variant="outlined" />
-          )
-        }}
-      />
-      <GenericForm {...{ ...props, fields: PartnerFormFields.BillingPartnerInformationFormFields }} />
+              <AccordionSummary
+                expandIcon={<ArrowForwardIosSharpIcon sx={{ fontSize: '0.9rem', color: '#536175' }} />}
+                aria-controls={`${formGroup.title}-content`}
+                id={`${formGroup.title}-content`}
+                sx={{
+                  py: 1,
+                  flexDirection: 'row-reverse',
+                  color: '#536175',
+                  transitionDuration: '500ms',
+                  "&:hover": {
+                    color: "#003133",
+                  },
+                  '& .MuiAccordionSummary-expandIconWrapper.Mui-expanded': {
+                    transform: 'rotate(90deg)',
+                  },
+                }}
+              >
+                <div className="pl-4">{formGroup.icon}</div>
+                <div className="pl-2 text-md font-medium">{formGroup.title}</div>
+              </AccordionSummary>
+              <AccordionDetails>
+                {
+                  formGroup.isCustomer ?
+                    <CustomerForm {...{ ...props, fields: formGroup.fields, customerType: formGroup.customerType }} /> :
+                    <GenericForm {...{ ...props, fields: formGroup.fields }} />
+                }
+
+              </AccordionDetails>
+            </Accordion>
+          ))
+        }
+      </div>
     </div>
   );
 };
